requirements.txt
langchain>=0.1.0
langgraph>=0.0.40
chromadb>=0.4.0
openai>=1.0.0
streamlit>=1.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
